feat(skill) : add Skill
오버래피드 제외 타 크로스보우 스킬 구현 완료

+++ Assets/Scripts/EqupimentManager/Weapon/WeaponSpec.cs
@@ -49,8 +49,8 @@ public class WeaponSpec : EquipSpec
     }
 
     [SerializeField]
-    private float weaponDamage;
-    public float WeaponDamage
+    private int weaponDamage;
+    public int WeaponDamage
     {
         get { return weaponDamage; }
         set { weaponDamage = value; }
